requirements.txt
discord.py-self==1.9.2
numpy
opencv-python-headless
pytesseract
requests
pillow
thefuzz
python-Levenshtein